Changed the satellite info and added saving
You can now save manually in the file menu, the program also asks you to
save when you close the window. It saves the hostIP, the port and your
site location.

The satellite info now includes the international designator, the apogee
height, the perigee height and the inclination of the satellites orbit.